+++ src/sphinx_codeautolink/extension/__init__.py
@@ -136,7 +136,7 @@ class SphinxCodeAutoLink:
         for doc, transforms in self.cache.transforms.items():
             if not transforms or str(Path(doc)) not in self.outdated_docs:
                 continue
-            file = (Path(app.outdir) / doc).with_suffix('.html')
+            file = Path(app.outdir) / (doc + '.html')
             link_html(file, app.outdir, transforms, self.make_inventory(app))
 
         self.cache.write()