+++ docs/src/release_notes.rst
@@ -20,6 +20,7 @@ Please delete the cache file before building documentation.
 - Fix crash on annotation-only assignment (:issue:`50`)
 - Fix issue with filenames that have dots (:issue:`52`)
 - Correctly remove extension when building non-HTML documentation (:issue:`53`)
+- Support searching extra CSS classes for code example matching (:issue:`54`)
 - Add configuration for global default concatenation state (:issue:`55`)
 
 0.4.0 (2021-10-08)